[CDSemanticString] Use array instead of linked list for storage
Effectively reverts commit 6f1dccc0371db322370d622cc6e1587d3e9bed27

In the append method of CDSemanticString, the nodes from the
"small" linked list are directly added to the list of the receiver.
This behavior means that a subsequent call to "append" on either
linked list would effectively change both lists. Fixing this
would require copying the nodes on append, which largely mitigates
the performance benefits of using a linked list here.